Renamed resource to cart instead of session
